refactoring: new folder cards in Client
